autoprov/cluster-svc package move fixes
