+++ cmd/autoprov/autoprov_main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/edgexr/edge-cloud-platform/pkg/autoprov"
+)
+
+func main() {
+	autoprov.Run()
+}
