+++ cmd/cluster-svc/cluster-svc-main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	clustersvc "github.com/edgexr/edge-cloud-platform/pkg/cluster-svc"
+)
+
+func main() {
+	clustersvc.Run()
+}
